+++ comnetsemu/net.py
@@ -21,7 +21,7 @@ from mininet.term import cleanUpScreens, makeTerms
 from mininet.util import BaseString
 
 # ComNetsEmu version: should be consistent with README and LICENSE
-VERSION = "0.1.7"
+VERSION = "0.1.8"
 
 APPCONTAINERMANGER_MOUNTED_DIR = "/tmp/comnetsemu/appcontainermanger"